Replace encrypt with rotate
Replaced the encrypt functions and their files with the equivalent
rotate counterpart and implemented rotate_char.

The rotate functions contain an additional parameter, align. Align is
the direction which the input should be rotated, allowing the one set of
functions to be used to encrypt and decrypt input.

Additionally, tied together the main function, and added comments to
the character set and binary search header files.

+++ src/char_set.h
@@ -1,8 +1,27 @@
 #include "bin_search.h"
 
-#define CHARSETLEN	'~' - ' ' + 2
+#define CHARSETLEN	(int)('~' - ' ' + 2)
 
+/**
+ * Define an array and fill it with the program's character set.
+ * 
+ * @param x Name of the array.
+ */
 #define get_char_set(x) char (x)[CHARSETLEN]; create_char_set((x));
+
+/**
+ * Search for the given value in the array.
+ * 
+ * @param x Array to search through.
+ * @param y Target value.
+ * 
+ * @return Index of target if found, else -1.
+ */
 #define get_char_index(x, y) (char)(bin_search((x), (y), 0, CHARSETLEN-1))
 
+/** 
+ * Fill an array with the program's character set.
+ * 
+ * @param char_set Array to fill.
+ */
 void create_char_set(char *char_set);
